+++ src/lib/vm/objects/instance.rs
@@ -1,0 +1,68 @@
+// Copyright (c) 2019 King's College London created by the Software Development Team
+// <http://soft-dev.org/>
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0>, or the MIT license <LICENSE-MIT or
+// http://opensource.org/licenses/MIT>, or the UPL-1.0 license <http://opensource.org/licenses/UPL>
+// at your option. This file may not be copied, modified, or distributed except according to those
+// terms.
+
+#![allow(clippy::new_ret_no_self)]
+
+use std::cell::UnsafeCell;
+
+use abgc_derive::GcLayout;
+
+use crate::vm::{
+    objects::{Class, Obj, ObjType, StaticObjType},
+    val::{NotUnboxable, Val},
+    vm::VM,
+};
+
+/// An instance of a user class.
+#[derive(Debug, GcLayout)]
+pub struct Inst {
+    class: Val,
+    inst_vars: UnsafeCell<Vec<Val>>,
+}
+
+impl Obj for Inst {
+    fn dyn_objtype(&self) -> ObjType {
+        ObjType::Inst
+    }
+
+    fn get_class(&self, _: &VM) -> Val {
+        self.class.clone()
+    }
+}
+
+impl NotUnboxable for Inst {}
+
+impl StaticObjType for Inst {
+    fn static_objtype() -> ObjType {
+        ObjType::Inst
+    }
+}
+
+impl Inst {
+    pub fn new(vm: &VM, class: Val) -> Val {
+        let cls: &Class = class.downcast(vm).unwrap();
+        let mut inst_vars = Vec::with_capacity(cls.num_inst_vars);
+        inst_vars.resize(cls.num_inst_vars, Val::illegal());
+        let inst = Inst {
+            class,
+            inst_vars: UnsafeCell::new(inst_vars),
+        };
+        Val::from_obj(vm, inst)
+    }
+
+    pub fn inst_var_lookup(&self, n: usize) -> Val {
+        let inst_vars = unsafe { &mut *self.inst_vars.get() };
+        inst_vars[n].clone()
+    }
+
+    pub fn inst_var_set(&self, n: usize, v: Val) {
+        let inst_vars = unsafe { &mut *self.inst_vars.get() };
+        inst_vars[n] = v;
+    }
+}
